api-data-send-model/requirements.txt
requests>=2.32.0
PyYAML>=6.0.1
tenacity>=9.0.0
pydantic>=2.8.0

